data-backup

+++ ppp-patrol/common/timeout_archive_clean.py
@@ -9,8 +9,8 @@ import _config
 import subprocess
 
 parser = argparse.ArgumentParser(description='Clean a folder')
-parser.add_argument('folder_path', type=str, help='The folder to clean')
-parser.add_argument('s3_path', type=str, help='The S3 path to upload to')
+parser.add_argument('--folder_path', type=str, help='The folder to clean')
+parser.add_argument('--s3_path', type=str, nargs='?', help='The S3 path to upload to')
 args = parser.parse_args()
 
 GZIP_TMP = '/tmp/XZRTVJAEZFPKYWTTXDESUPEAYLSLNLUP'
@@ -41,14 +41,15 @@ for file_path in file_path_list:
 
     basename = os.path.basename(file_path)
 
-    print('Compressing...')
-    subprocess.run(['gzip', '-c', file_path], stdout=open(GZIP_TMP, 'wb'), check=True)
+    if s3_path != None:
+        print('Compressing...')
+        subprocess.run(['gzip', '-c', file_path], stdout=open(GZIP_TMP, 'wb'), check=True)
 
-    print('Uploading to S3...')
-    s3.upload_file(GZIP_TMP, _config.S3_BUCKET, f'archive/{s3_path}/{file_mtime_yyyy}/{file_mtime_mm}/{basename}.gz')
+        print('Uploading to S3...')
+        s3.upload_file(GZIP_TMP, _config.S3_BUCKET, f'archive/{s3_path}/{file_mtime_yyyy}/{file_mtime_mm}/{basename}.gz')
 
-    print('Removing temporary file...')
-    os.remove(GZIP_TMP)
+        print('Removing temporary file...')
+        os.remove(GZIP_TMP)
 
     print('Removing local file...')
     os.remove(file_path)
